requirements-test.txt
mock==2.0.0
vcrpy==1.6.1